admin control only in users table

+++ resources/views/livewire/components/users-table.blade.php
@@ -27,12 +27,14 @@
             <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
               Localisation
             </th>
-            <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
-              Rôle
-            </th>
-            <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
-              Actions
-            </th>
+            @if (auth()->user()->isAdmin())
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                Rôle
+              </th>
+              <th class="text-center text-uppercase text-secondary text-xxs font-weight-bolder opacity-7">
+                Actions
+              </th>
+            @endif
           </tr>
         </thead>
         <tbody>
@@ -50,17 +52,19 @@
               <td class="text-md-center" style="cursor: pointer" wire:click='selectEvent({{ $user->id }})'>
                 <p class="text-xs font-weight-bold mb-0 ps-3">{{ $user->location }}</p>
               </td>
-              <td class="text-md-center" style="cursor: pointer" wire:click='selectEvent({{ $user->id }})'>
-                <p class="text-xs font-weight-bold mb-0 ps-3">{{ $user->role }}</p>
-              </td>
-              <td class="text-md-center">
-                <a href="{{ route('users.show', [$user->id]) }}" class="mx-3" data-bs-toggle="tooltip" data-bs-original-title="Editer l'utilisateur">
-                  <i class="fas fa-user-edit text-secondary"></i>
-                </a>
-                <span>
-                  <i class="cursor-pointer fas fa-trash text-secondary" data-bs-toggle="tooltip" data-bs-original-title="Supprimer l'utilisateur" onclick="confirm('Supprimer l\'utilisateur : {{ $user->name }} ?') || event.stopImmediatePropagation()" wire:click="deleteUser({{ $user->id }})"></i>
-                </span>
-              </td>
+              @if (auth()->user()->isAdmin())
+                <td class="text-md-center" style="cursor: pointer" wire:click='selectEvent({{ $user->id }})'>
+                  <p class="text-xs font-weight-bold mb-0 ps-3">{{ $user->role }}</p>
+                </td>
+                <td class="text-md-center">
+                  <a href="{{ route('users.show', [$user->id]) }}" class="mx-3" data-bs-toggle="tooltip" data-bs-original-title="Editer l'utilisateur">
+                    <i class="fas fa-user-edit text-secondary"></i>
+                  </a>
+                  <span>
+                    <i class="cursor-pointer fas fa-trash text-secondary" data-bs-toggle="tooltip" data-bs-original-title="Supprimer l'utilisateur" onclick="confirm('Supprimer l\'utilisateur : {{ $user->name }} ?') || event.stopImmediatePropagation()" wire:click="deleteUser({{ $user->id }})"></i>
+                  </span>
+                </td>
+              @endif
             </tr>
           @endforeach
         </tbody>
